reorganize preparedness caching
