+++ plugins/polio/api.py
@@ -2,7 +2,7 @@ import csv
 import functools
 import json
 import datetime as dt
-from datetime import timedelta, datetime, date
+from datetime import timedelta, datetime
 from typing import Any, Optional
 from collections import defaultdict
 from functools import lru_cache
@@ -57,15 +57,13 @@ from .models import (
     Campaign,
     Config,
     LineListImport,
-    SpreadSheetImport,
     Round,
     CampaignGroup,
     CampaignScope,
     RoundScope,
 )
 from .models import CountryUsersGroup
-from .preparedness.calculator import preparedness_summary
-from .preparedness.parser import get_preparedness
+from .preparedness.calculator import get_or_set_preparedness_cache_for_round
 from .export_utils import generate_xlsx_campaigns_calendar, xlsx_file_name
 
 logger = getLogger(__name__)
@@ -441,6 +439,7 @@ Timeline tracker Automated message
         if cached_response:
             return cached_response
 
+        # noinspection SqlResolve
         round_scope_queryset = campaigns.filter(separate_scopes_per_round=True).annotate(
             geom=RawSQL(
                 """select st_asgeojson(st_simplify(st_union(st_buffer(iaso_orgunit.simplified_geom::geometry, 0)), 0.01)::geography)
@@ -453,6 +452,7 @@ where polio_round.campaign_id = polio_campaign.id""",
             )
         )
         # For campaign scope
+        # noinspection SqlResolve
         campain_scope_queryset = campaigns.filter(separate_scopes_per_round=False).annotate(
             geom=RawSQL(
                 """select st_asgeojson(st_simplify(st_union(st_buffer(iaso_orgunit.simplified_geom::geometry, 0)), 0.01)::geography)
@@ -563,8 +563,8 @@ where group_id = polio_roundscope.group_id""",
             )
         )
 
+        scope: RoundScope
         for scope in round_scopes:
-            scope: RoundScope
             if scope.geom:
                 feature = {
                     "type": "Feature",
@@ -632,89 +632,6 @@ class LineListImportViewSet(ModelViewSet):
         return LineListImport.objects.all()
 
 
-DAYS_EVOLUTION = [
-    # day before, target in percent
-    (1, 90),
-    (3, 85),
-    (7, 80),
-    (14, 60),
-    (21, 40),
-    (28, 20),
-]
-
-
-def score_for_x_day_before(ssi_for_campaign, ref_date: date, n_day: int):
-    day = ref_date - timedelta(days=n_day)
-    try:
-        ssi = ssi_for_campaign.filter(created_at__date=day).last()
-    except SpreadSheetImport.DoesNotExist:
-        return None, day, None
-    try:
-        preparedness = get_preparedness(ssi.cached_spreadsheet)
-        summary = preparedness_summary(preparedness)
-        score = summary["overall_status_score"]
-    except Exception as e:
-        return None, day, None
-    return ssi.created_at, day, score
-
-
-def history_for_campaign(ssi_qs, round: Round):
-    ref_date = round.started_at
-    if not ref_date:
-        return {"error": f"Please configure a start date for the round {round}"}
-    r = []
-    for n_day, target in DAYS_EVOLUTION:
-        sync_time, day, score = score_for_x_day_before(ssi_qs, ref_date, n_day)
-        r.append(
-            {
-                "days_before": n_day,
-                "expected_score": target,
-                "preparedness_score": score,
-                "date": day,
-                "sync_time": sync_time,
-            }
-        )
-    return r
-
-
-def _make_prep(c: Campaign, round: Round):
-    logger.info(f"Make prep called for round {round} - {c}")
-    url = round.preparedness_spreadsheet_url
-    if not url:
-        return None
-    campaign_prep = {
-        "campaign_id": c.id,
-        "campaign_obr_name": c.obr_name,
-        "indicators": {},
-        "round": f"Round{round.number}",
-        "round_id": round.id,
-        "round_start": round.started_at,
-        "round_end": round.ended_at,
-    }
-    try:
-        spread_id = extract_id_from_url(url)
-        ssi_qs = SpreadSheetImport.objects.filter(spread_id=spread_id)
-        last_ssi = ssi_qs.last()
-        if not ssi_qs or not last_ssi:
-            # No import yet
-            campaign_prep["status"] = "not_sync"
-            campaign_prep["details"] = "This spreadsheet has not been synchronised yet"
-            return campaign_prep
-        campaign_prep["date"] = last_ssi.created_at
-        cs = last_ssi.cached_spreadsheet
-        last_p = get_preparedness(cs)
-        campaign_prep.update(preparedness_summary(last_p))
-        if round.number != last_p["national"]["round"]:
-            logger.info(f"Round mismatch on {c} {round}")
-
-        campaign_prep["history"] = history_for_campaign(ssi_qs, round)
-    except Exception as e:  # FIXME: too broad Exception
-        campaign_prep["status"] = "error"
-        campaign_prep["details"] = str(e)
-        logger.exception(e)
-    return campaign_prep
-
-
 class PreparednessDashboardViewSet(viewsets.ViewSet):
     def list(self, request):
 
@@ -725,13 +642,7 @@ class PreparednessDashboardViewSet(viewsets.ViewSet):
 
         for c in qs:
             for round in c.rounds.all():
-                cache: BaseCache
-                p = cache.get_or_set(
-                    f"prepardness-{round.id}",
-                    default=lambda: _make_prep(c, round),
-                    timeout=60 * 60 * 8,  # eighth hours
-                )
-
+                p = get_or_set_preparedness_cache_for_round(c, round)
                 if p:
                     r.append(p)
         return Response(r)
