+++ plugins/polio/management/commands/refresh_preparedness_data.py
@@ -1,11 +1,10 @@
 from datetime import datetime
 from logging import getLogger
 
-from django.core.cache import cache
 from django.core.management.base import BaseCommand
 
-from plugins.polio.api import _make_prep
 from plugins.polio.models import Campaign, SpreadSheetImport, Round
+from plugins.polio.preparedness.calculator import set_preparedness_cache_for_round
 
 logger = getLogger(__name__)
 
@@ -33,11 +32,8 @@ class Command(BaseCommand):
                 logger.info(f"using spread: {cs.title}")
                 round.preparedness_sync_status = "FINISHED"
                 round.save()
-                cache.set(
-                    f"prepardeness-{round.id}",
-                    value=_make_prep(round.campaign, round),
-                    timeout=60 * 60 * 24,  # 24 hours
-                )
+                set_preparedness_cache_for_round(round)
+
             except Exception as e:
                 logger.error(f"Round {round} refresh failed")
                 logger.exception(e)
